backend/requirements.txt
# FastAPI and web server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# HTTP and HTML parsing
httpx==0.28.1
beautifulsoup4==4.12.3
selectolax==0.3.24

# Utilities
pydantic==2.10.5
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pyyaml==6.0.2

# AI/LLM
google-genai==0.2.2

# Playwright for robust scraping of dynamic pages (RCMP parser)
playwright

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==4.1.0
